refactor [main.js]: add async syntax
